Fix some warnings occuring on FreeBSD

+++ src/lib/libast/misc/getcwd.c
@@ -81,7 +81,7 @@ getcwd(char* buf, size_t len)
 struct dirlist				/* long path chdir(2) component	*/
 {
 	struct dirlist*	next;		/* next component		*/
-	int		index;		/* index from end of buf	*/
+	ssize_t		index;		/* index from end of buf	*/
 };
 
 /*
@@ -147,8 +147,8 @@ getcwd(char* buf, size_t len)
 	char*		p;
 	char*		s;
 	DIR*		dirp = 0;
-	int		n;
-	int		x;
+	size_t		n;
+	ssize_t		x;
 	size_t		namlen;
 	ssize_t		extra = -1;
 	struct dirent*	entry;
@@ -271,7 +271,7 @@ getcwd(char* buf, size_t len)
 		{
 			if (!(entry = readdir(dirp))) ERROR(ENOENT);
 			namlen = D_NAMLEN(entry);
-			if ((d - dots) > (PATH_MAX - 1 - namlen))
+			if ((d - dots) > (ssize_t)(PATH_MAX - 1 - namlen))
 			{
 				*d = 0;
 				if (namlen >= PATH_MAX || !(dirstk = pushdir(dirstk, dots + 3, p, buf + len - 1))) ERROR(ERANGE);
